Create SaveFs class
